AULA04/Tarefa4/GatewaySolution/CommandModule/requirements.txt
smbus2
flask
blynk-library-python